requirements.txt
django==4.0.6
psycopg2-binary==2.9.3
requests==2.28.1
pillow==9.2.0
djangorestframework==3.13.1
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-filter==22.1
django-rq==2.5.1
pytest>=7.1.2
faker>=13.15.0
pytest-django>=4.5.2
factory-boy>=3.2.1
pytest-factoryboy==2.5.0
redis==4.3.4
Scrapy==2.6.2
python-telegram-bot==20.0a4
gunicorn==20.1.0

flake8>=5.0.4
isort>=5.10.1
black>=22.10.0

